+++ Editor/CoverageFormats/OpenCover/OpenCoverResultWriter.cs
@@ -2,7 +2,6 @@ using System.Xml.Serialization;
 using System.IO;
 using UnityEngine;
 using OpenCover.Framework.Model;
-using UnityEditor.TestTools.CodeCoverage.Utils;
 
 namespace UnityEditor.TestTools.CodeCoverage.OpenCover
 {